fix: Refactor docs
