+++ modules/react/icon/stories/IconsList.stories.mdx
@@ -1,16 +1,11 @@
-import {SystemIconList} from './examples/IconList';
-import {AppletIconList} from './examples/AppletIconList';
-import {AccentIconList} from './examples/AccentIconList.tsx';
-import {Overview} from './examples/Overview';
-import {Images} from './examples/Image';
+import {SystemIconList} from './examples/IconsList/IconList';
+import {AppletIconList} from './examples/IconsList/AppletIconList';
+import {AccentIconList} from './examples/IconsList/AccentIconList';
+import {Overview} from './examples/Overview'
 
-<Meta title="Assets/Icons" />
+<Meta title="Images and Icons/Icons List" />
 
-# Assets
-
-foo bar Assets are graphics which help communicate meaning or highlight areas of interaction to our
-users. These may be presented individually or as part of a group with related information in
-components and patterns.
+## Overview
 
 <ExampleCodeBlock code={Overview} />
 
@@ -34,7 +29,3 @@ System Icons are symbols used to convey simple actions and functions, they are t
 encountered in products and help communicate metaphors at a glance.
 
 <ExampleCodeBlock code={SystemIconList} />
-
-## Images
-
-<ExampleCodeBlock code={Images} />
