+++ modules/react/icon/stories/Image.stories.mdx
@@ -1,0 +1,8 @@
+
+import {Basic} from './examples/Image/Basic';
+
+<Meta title="Images and Icons/Image" />
+
+## Images
+
+<ExampleCodeBlock code={Basic} />
